+++ gamemodes/terrortown/entities/weapons/weapon_ttt_beacon.lua
@@ -8,13 +8,18 @@ end
 
 DEFINE_BASECLASS("weapon_tttbase")
 
-SWEP.HoldType = "normal"
+SWEP.HoldType = "slam"
 
 if CLIENT then
     SWEP.PrintName = "beacon_name"
     SWEP.Slot = 6
 
+    SWEP.ViewModelFOV = 70
+    SWEP.ViewModelFlip = false
+
+    SWEP.UseHands = true
     SWEP.ShowDefaultViewModel = false
+    SWEP.ShowDefaultWorldModel = false
 
     SWEP.EquipMenuData = {
         type = "item_weapon",
@@ -26,8 +31,8 @@ end
 
 SWEP.Base = "weapon_tttbase"
 
-SWEP.ViewModel = "models/weapons/v_crowbar.mdl"
-SWEP.WorldModel = "models/props_lab/reciever01b.mdl"
+SWEP.ViewModel = "models/weapons/cstrike/c_c4.mdl"
+SWEP.WorldModel = "models/props_lab/reciever01a.mdl"
 
 SWEP.Primary.ClipSize = 3
 SWEP.Primary.DefaultClip = 1
@@ -117,26 +122,43 @@ function SWEP:Reload()
     return false
 end
 
----
--- @realm shared
-function SWEP:OnRemove()
-    if
-        CLIENT
-        and IsValid(self:GetOwner())
-        and self:GetOwner() == LocalPlayer()
-        and self:GetOwner():IsTerror()
-    then
-        RunConsoleCommand("lastinv")
-    end
-end
-
 if CLIENT then
     ---
     -- @realm client
     function SWEP:Initialize()
         self:AddTTT2HUDHelp("beacon_help_pri", "beacon_help_sec")
 
-        return BaseClass.Initialize(self)
+        self:AddCustomViewModel("vmodel", {
+            type = "Model",
+            model = "models/props_lab/reciever01a.mdl",
+            bone = "ValveBiped.Bip01_R_Finger2",
+            rel = "",
+            pos = Vector(2.2, 6.5, -1),
+            angle = Angle(120, 10, 0),
+            size = Vector(0.6, 0.6, 0.6),
+            color = Color(255, 255, 255, 255),
+            surpresslightning = false,
+            material = "",
+            skin = 0,
+            bodygroup = {},
+        })
+
+        self:AddCustomWorldModel("wmodel", {
+            type = "Model",
+            model = "models/props_lab/reciever01a.mdl",
+            bone = "ValveBiped.Bip01_R_Hand",
+            rel = "",
+            pos = Vector(6.7, 7, -1),
+            angle = Angle(-60, 35, 0),
+            size = Vector(0.6, 0.6, 0.6),
+            color = Color(255, 255, 255, 255),
+            surpresslightning = false,
+            material = "",
+            skin = 0,
+            bodygroup = {},
+        })
+
+        self.BaseClass.Initialize(self)
     end
 
     ---
@@ -148,18 +170,4 @@ if CLIENT then
             RunConsoleCommand("lastinv")
         end
     end
-
-    ---
-    -- @realm client
-    function SWEP:DrawWorldModel()
-        if IsValid(self:GetOwner()) then
-            return
-        end
-
-        self:DrawModel()
-    end
-
-    ---
-    -- @realm client
-    function SWEP:DrawWorldModelTranslucent() end
 end